Remove "->" from the end of variables detected as class instances
The added "->" only helps you when you want to call a method on that
variable, but there are scenarios when you want to use just the variable,
like passing it to a function as parameter or using it in a string
concatenation / interpolation. Half of the times you will end up deleting it.