Replace normal! % to a searchpair() call
For some reason, the normal! with the bang(!) does not work in some cases
where the normal (without bang) or searchpair() does. Since the % movement
is usually enhanced with plugins and it's behavior is configuration
dependent it's better to use searchpair() where we can control it's
behavior more closely.
I've found this problem during the debug of #35, thanks again for the
testing @markwu