Implement a smarter <C-]>
The plugin will try to guess the expression's class under cursor and jump
to the correct one instead of just matching the tagnames. Falls back to
the built-in method on failure.
The enhancement can be disabled with the new
g:phpcomplete_enhance_jump_to_definition option

This feature was first mentioned in #31.