Enhanced jump redux from feedkeys() to :tag
Now the phpcomplete#JumpToDefinition() command will use the :tag command
to the actual jumping instead of feedkeys(). This is much more robust and
does not require temp files or modifying the tags settings.

Special thanks to @dsawardekar for the wordpress.vim plugin where I've
stolen the idea from.

Related to #44