Change jump_to_def_vsplit default to <C-W>}
The previous <C-W><C-[> turns out to be a bad idea since <C-[> usually
means <Esc> in readline and other places.
See: https://en.wikipedia.org/wiki/Control_key
On top of that it seem to cause issues with the vim-session plugin, thanks
for @paprykarz for the investigation.
Discussion at issue: #51