phpcomplete#GetClassName(): Avoid infinite loop
Because of too much simple regexp to detect classes
it's not hard to get GetClassName() iterating forever.

This commit just adds a simple \s to the line regexp in
order to avoid some obvious false matches. And then moves
the inc to the end of the loop, so it will
happen always, ensuring the infinite loop won't happen ever.

I've tried also adding some "forbidden" chars to the line
regexp, but finally decided to abandon that route because
realized that somebody could be adding those chars in a comment
or other endless combinations.

Simple code examples heading to the infinite loop:

$var = 'Anything with a
       class will loop to the infinite'

$verylongvariablethatcausesnextlinetowrap =
    classname(...);

classify_is_a_nice_method_name_isnt_it(...);

And so on. This simple fix should avoid the problem.