Add some smoke tests related to #62
While the plugin can still be fooled with strings the added test would have
ran forever without the patch in #62. The same test could be used later
with the proper values in the assertion to test that if we are matching
the "class" keyword inside a string or comment or not.