Update and fix openssl_*()
This updates and fixes definitions for many OpenSSL-related built-in
functions.

Fixed the parameter names to match the documentation (this is to aid
users of PHP 8's named parameters).

Use class names as types instead of "resource" where appropriate; in
cases where it was two or more types, it was left as "mixed".

Fix incorrect signature for `openssl_pkcs7_encrypt`.

Minor fixes to return type; true/false/-1 on error => bool.

Function support added:

* [openssl\_cipher\_key\_length](https://www.php.net/manual/en/function.openssl-cipher-key-length.php) (added in PHP 8.2.0)
* [openssl\_cms\_decrypt](https://www.php.net/manual/en/function.openssl-cms-decrypt.php) (added in PHP 8)
* [openssl\_cms\_encrypt](https://www.php.net/manual/en/function.openssl-cms-encrypt.php) (added in PHP 8)
* [openssl\_cms\_sign](https://www.php.net/manual/en/function.openssl-cms-sign.php) (added in PHP 8)
* [openssl\_cms\_verify](https://www.php.net/manual/en/function.openssl-cms-verify.php) (added in PHP 8)
* [openssl\_pkey\_derive](https://www.php.net/manual/en/function.openssl-pkey-derive.php) (added in PHP 7.3.0)
* [openssl\_x509\_verify](https://www.php.net/manual/en/function.openssl-x509-verify.php) (added in PHP 7.4.0)